+++ resources/views/questions/index.blade.php
@@ -35,10 +35,16 @@
                                     <div class="d-flex align-items-center">
                                         <h3 class="mt-0"><a href="{{ $question->url }}">{{ $question->title }}</a></h3>
                                         <div class="ml-auto">
-                                            <a href="{{ route('questions.edit', $question->id) }}" class="btn btn-sm btn-outline-info">Edit</a>
-                                            {!! Form::open(['method' => 'DELETE', 'action' => ['QuestionController@destroy', $question->id],'class' => 'form-delete']) !!}
+                                            @if(Auth::user()->can('update-question', $question))
+                                                <a href="{{ route('questions.edit', $question->id) }}"
+                                                   class="btn btn-sm btn-outline-info">Edit</a>
+                                            @endif
+
+                                            @if(Auth::user()->can('delete-question', $question))
+                                                {!! Form::open(['method' => 'DELETE', 'action' => ['QuestionController@destroy', $question->id],'class' => 'form-delete']) !!}
                                                 {!! Form::submit('Delete', ['class' =>'btn btn-outline-danger btn-sm', 'onclick' => 'return confirm("Are you sure?")']); !!}
-                                            {!! Form::close() !!}
+                                                {!! Form::close() !!}
+                                            @endif
                                         </div>
                                     </div>
                                     <p class="lead">
